pyproject.toml
[project]
name = "adk-playwright-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.4.1",
]
